Compilation database support (#22)
* Switch to Box<dyn ...> over Box<...>

(The latter was deprecated.)

* Add basic compile command database bindings.

Clang provides APIs for working with a compilation database (a
`compile_commands.json` file that maps each compiled source file to
the command and args used to build it). Most of the needed functions
are already in clang-sys, but clang-rs provided no high-level binds.

This adds some fairly direct bindings to the C API:
from a directory, you get a `CompilationDatabase` struct. From there you
can get a `CompileCommands` struct that represents all commands, or ones
for a provided path. From that, you finally get a `Vec<CompileCommand>`,
each of which provides its working directory, filename, args, etc.

This seems a little clunky - it would be nice for the
`CompilationDatabase` calls to return a `Vec<CompileCommand>` directly,
but the C API implies that the data for each individual `CompileCommand`
is tied to the `CompileCommands` it came from.